Fix linter file
